Consolidate `test-utils.tsx` files
