+++ clients/admin-ui/__tests__/nav.test.tsx
@@ -1,9 +1,8 @@
 import { SessionProvider } from "next-auth/react";
 
 import NavBar from "~/features/common/NavBar";
-import { mockNextUseRouter } from "~/test-utils";
 
-import { render } from "./test-utils";
+import { mockNextUseRouter, render } from "./test-utils";
 
 describe("Home", () => {
   it("renders a navigation bar", () => {
