+++ clients/admin-ui/__tests__/test-utils.tsx
@@ -4,7 +4,7 @@ import { render as rtlRender, RenderOptions } from "@testing-library/react";
 import React from "react";
 import { Provider } from "react-redux";
 
-import { makeStore, RootState } from "../src/app/store";
+import { makeStore, RootState } from "~/app/store";
 
 type CustomRenderOptions = {
   preloadedState?: Partial<RootState>;
@@ -30,3 +30,30 @@ export * from "@testing-library/react";
 
 // override render method
 export { render };
+
+// Mocks useRouter
+const useRouter = jest.spyOn(require("next/router"), "useRouter");
+
+/**
+ * mockNextUseRouter
+ * Mocks the useRouter React hook from Next.js on a test-case by test-case basis
+ * Adapted from https://github.com/vercel/next.js/issues/7479#issuecomment-587145429
+ */
+export function mockNextUseRouter({
+  route = "/",
+  pathname = "/",
+  query = "/",
+  asPath = "/",
+}: {
+  route?: string;
+  pathname?: string;
+  query?: string;
+  asPath?: string;
+}) {
+  useRouter.mockImplementation(() => ({
+    route,
+    pathname,
+    query,
+    asPath,
+  }));
+}
